requirements.txt
Django>=5.2.5
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
djangorestframework-simplejwt>=5.3.0